+++ src/types/index.ts
@@ -4,7 +4,7 @@ export * from "./models/Cheerio.ts";
 export * from "./models/Puppeteer.ts";
 export * from "./models/Context.ts";
 export * from "./models/Xhr.ts";
-export { ProvidersManifest, GithubSource, RegistrySource, LocalSource, ProviderSource, ProviderManagerConfig } from "./models/Manager.ts";
+export { ProvidersManifest, ExternalProviderManifest, GithubSource, RegistrySource, LocalSource, ProviderSource, ProviderManagerConfig } from "./models/Manager.ts";
 export * from "./models/Modules.ts";
 
 // Input and Output Types